🚚Rename: data 폴더 이동 후 경로 수정
